+++ .project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>Ejercicios-en-repositorio</name>
+	<name>marcinikDawid</name>
 	<comment></comment>
 	<projects>
 	</projects>
