--- src/junit/ej2/FechaException.java
@@ -1,14 +1,0 @@
-package junit.ej2;
-
-public class FechaException extends Exception{
-
-	public FechaException(String message) {
-		super(message);
-	}
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-}
