--- src/junit/ej3/NumeroException.java
@@ -1,14 +1,0 @@
-package junit.ej3;
-
-public class NumeroException extends Exception{
-
-	public NumeroException(String message) {
-		super(message);
-	}
-	
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-}
