--- src/junit/ej4/InputDataException.java
@@ -1,13 +1,0 @@
-package junit.ej4;
-
-public class InputDataException extends Exception{
-
-	public InputDataException(String message) {
-		super(message);
-	}
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-}
