+++ src/org/marcinikDawid/tema8/ej02/FechaException.java
@@ -1,0 +1,14 @@
+package org.marcinikDawid.tema8.ej02;
+
+public class FechaException extends Exception{
+
+	public FechaException(String message) {
+		super(message);
+	}
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+
+}
