+++ src/org/marcinikDawid/tema8/ej03/NumeroException.java
@@ -1,0 +1,14 @@
+package org.marcinikDawid.tema8.ej03;
+
+public class NumeroException extends Exception{
+
+	public NumeroException(String message) {
+		super(message);
+	}
+	
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+
+}
