+++ src/org/marcinikDawid/tema8/ej04/InputDataException.java
@@ -1,0 +1,13 @@
+package org.marcinikDawid.tema8.ej04;
+
+public class InputDataException extends Exception{
+
+	public InputDataException(String message) {
+		super(message);
+	}
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+
+}
